back_end_django/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
drf-yasg
python-decouple
psycopg2-binary
djangorestframework-simplejwt